rag_11/requirements.txt
annotated-types==0.7.0
anyio==4.12.0
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
distro==1.9.0
google-auth==2.45.0
google-genai==1.56.0
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==1.2.5
langchain-text-splitters==1.1.0
langsmith==0.5.1
numpy==2.4.0
orjson==3.11.5
packaging==25.0
portalocker==3.2.0
protobuf==6.33.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
pywin32==311
PyYAML==6.0.3
qdrant-client==1.16.2
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
sniffio==1.3.1
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uuid_utils==0.12.0
websockets==15.0.1
zstandard==0.25.0
